rename(bottomDialog): change name signatures(directory) - before: fragment - after: bottomDialog
